+++ AssemblyInfo.cs
@@ -6,6 +6,6 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyProduct("TrayMinimizer")]
 [assembly: AssemblyCompany("Brendon Kupsch")]
 [assembly: AssemblyCopyright("Copyright (c) 2026 Brendon Kupsch. MIT licensed.")]
-[assembly: AssemblyVersion("1.0.0.0")]
-[assembly: AssemblyFileVersion("1.0.0.0")]
+[assembly: AssemblyVersion("1.0.1.0")]
+[assembly: AssemblyFileVersion("1.0.1.0")]
 [assembly: ComVisible(false)]